Adds testmode in setup_experiment.
